FIX: Course 엔티티의 distance 필드 타입 변경 (#89)
코스의 distance 필드 타입을 int에서 double로 변경하였습니다.

+++ src/main/java/com/server/running_handai/domain/course/dto/CourseInfoWithDetailsDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record CourseInfoWithDetailsDto(
         long courseId,
         String thumbnailUrl,
-        int distance,
+        double distance,
         int duration,
         double maxElevation,
         double distanceFromUser,
